perf: stream cycle detail data with the RSC payload
The cycle detail page was fully client-rendered: the browser had to
download and parse JS, then fire cycle.getById + workspace.getBySlug,
then a second-tier product.getBySlug that waits on the workspace
result — three network hops after first paint before content appeared.

- Split the page into a server shell + CycleDetailClient and prefetch
  cycle.getById, workspace.getBySlug and product.product.getBySlug via
  HydrateClient, so all three queries stream with the initial document
  and the client renders without any post-load fetch round trips.
- Collapse cycle.getById from 3-4 sequential DB queries (existence
  pre-check, membership, full refetch) to a single fetch plus a
  membership check against the already-loaded row.
