+++ src/plugins/product/server/routers/cycle.ts
@@ -317,8 +317,9 @@ export const cycleRouter = createTRPCRouter({
   getById: protectedProcedure
     .input(z.object({ id: z.string() }))
     .query(async ({ ctx, input }) => {
-      await loadCycleWithAccess(ctx.db, ctx.session.user.id, input.id);
-
+      // Single fetch instead of loadCycleWithAccess + refetch: each DB round
+      // trip is sequential latency on the page's critical path, so assert
+      // access against the already-loaded row.
       const cycle = await ctx.db.list.findUnique({
         where: { id: input.id },
         include: {
@@ -348,6 +349,7 @@ export const cycleRouter = createTRPCRouter({
       if (!cycle || cycle.listType !== "SPRINT") {
         throw new TRPCError({ code: "NOT_FOUND", message: "Cycle not found" });
       }
+      await assertWorkspaceMember(ctx.db, ctx.session.user.id, cycle.workspaceId);
       return cycle;
     }),
